move impact cell view model to own file
